add more strings and move map to a different object
