+++ src/main/kotlin/crimera/patches/twitter/misc/bringbacktwitter/StringsMap.kt
@@ -1,0 +1,66 @@
+package crimera.patches.twitter.misc.bringbacktwitter
+
+object StringsMap {
+    val replacementMap = mapOf(
+        "button_action_add_tweet" to "Add a tweet",
+        "button_action_options_tweet" to "Tweet options",
+        "button_action_retweet" to "Retweet",
+        "button_new_tweet" to "New tweet",
+        "button_status_retweeted" to "Retweet (retweeted)",
+        "chrome_action_post" to "Tweet",
+        "ps__share_post_tweet" to "Tweet",
+        "rux_landing_page_title" to "Tweet",
+        "tweet_fab_item" to "Tweet",
+        "tweet_title" to "Tweet",
+        "search_twitter" to "Search Twitter",
+        "search_hint" to "Search Twitter",
+        "ps__accessibility_retweet_broadcast_button" to "Retweet",
+        "shortcut_retweet_tweet" to "Retweet",
+        "social_you_retweeted" to "You retweeted",
+        "composer_hint_self_thread" to "Add another tweet",
+        "confirm_delete_shared_tweet_description" to "Are you sure you want to remove this tweet?",
+        "confirm_delete_shared_tweet_title" to "Remove tweet",
+        "conversation_control_reply_restricted_dialog_title" to "Tweet not sent",
+        "conversation_control_reply_restricted_error" to "Tweet not sent. Replies are restricted.",
+        "conversations_alternative_reply_hint" to "Tweet your reply",
+        "conversations_other_tweets" to "Other tweets",
+        "curation_i_dont_like_this_tweet" to "I don’t like this tweet",
+        "date_posted" to "Date tweeted",
+        "post_tweet" to "Tweet",
+        "deleted_tweet_title" to "This tweet has been deleted.",
+        "dm_sensitive_tweet_interstitial_header" to "This tweet may contain sensitive material",
+        "dm_untrusted_tweet_interstitial_header" to "This tweet is hidden",
+        "empty_profile_tweets_tab_title" to "You haven’t tweeted yet",
+        "feedback_action_report_tweet" to "Report tweet",
+        "feedback_tweet_unavailable" to "This tweet is unavailable.",
+        "filter_item_tweets" to "Tweets",
+        "filter_item_tweets_and_replies" to "Tweets & replies",
+        "profile_tab_title_timeline" to "Tweets",
+        "users_turn_off_retweets" to "Turn off retweets",
+        "tweets_retweet" to "Retweet",
+        "tweets_retweeted" to "%s retweeted",
+        "tweets_undo_retweet_vertical" to "Undo retweet",
+        "users_turn_on_retweets" to "Turn on retweets",
+        "ps__post_broadcast_twitter" to "Tweet on Twitter",
+        "ps__retweet_broadcast_action" to "Retweet on Twitter",
+        "retweeters_title" to "Retweeted by",
+        "icon_view_tweet_activity" to "View tweet activity",
+        "a11y_views_text" to "View tweet activity",
+        "tweet_analytics_title" to "Tweet activity",
+        "tweets_delete_title" to "Delete tweet",
+        "share_tweet_sheet_title" to "Share tweet",
+        "view_quote_tweet" to "View tweet",
+        "view_tweet" to "View tweet",
+        "view_tweet_text" to "View tweet",
+        "quote_label_subtitle" to "Add a comment, photo, or GIF before you share this tweet",
+        "retweet_label_subtitle" to "Share this tweet with your followers",
+        "tweet_added_to_your_bookmarks" to "Tweet added to your Bookmarks",
+        "tweet_author" to "Tweet author",
+        "tweets_retweeted_accessibility_description" to "Tweeted by %s",
+        "tweets_topic_accessibility_description" to "Tweet from topic %s",
+        "tweets_unauthorized_error" to "Sorry, you’re not authorized to view these tweets.",
+        "ps__posted_on_twitter" to "*%s* tweeted on",
+        "ps__retweeted_on_twitter" to "*%s* retweeted on",
+        "tweet_removed_from_your_bookmarks" to "Tweet removed from your Bookmarks"
+    )
+}
